Extract workflow dispatch logic into reusable mediator command handlers
