+++ src/core/Elsa.Core/Dispatch/Consumers/ExecuteWorkflowInstanceRequestConsumer.cs
@@ -1,10 +1,8 @@
 using System.Diagnostics;
-using System.Linq;
 using System.Threading.Tasks;
 using Elsa.DistributedLocking;
-using Elsa.Models;
-using Elsa.Persistence;
 using Elsa.Services;
+using MediatR;
 using Microsoft.Extensions.Logging;
 using NodaTime;
 using Rebus.Handlers;
@@ -13,22 +11,19 @@ namespace Elsa.Dispatch.Consumers
 {
     public class ExecuteWorkflowInstanceRequestConsumer : IHandleMessages<ExecuteWorkflowInstanceRequest>
     {
-        private readonly IResumesWorkflow _workflowRunner;
-        private readonly IWorkflowInstanceStore _workflowInstanceStore;
+        private readonly IMediator _mediator;
         private readonly IDistributedLockProvider _distributedLockProvider;
         private readonly ICommandSender _commandSender;
         private readonly ILogger _logger;
         private readonly Stopwatch _stopwatch = new();
 
         public ExecuteWorkflowInstanceRequestConsumer(
-            IResumesWorkflow workflowRunner, 
-            IWorkflowInstanceStore workflowInstanceStore, 
+            IMediator mediator, 
             IDistributedLockProvider distributedLockProvider,
             ICommandSender commandSender,
             ILogger<ExecuteWorkflowInstanceRequestConsumer> logger)
         {
-            _workflowRunner = workflowRunner;
-            _workflowInstanceStore = workflowInstanceStore;
+            _mediator = mediator;
             _distributedLockProvider = distributedLockProvider;
             _commandSender = commandSender;
             _logger = logger;
@@ -51,15 +46,7 @@ namespace Elsa.Dispatch.Consumers
             
             try
             {
-                var workflowInstance = await _workflowInstanceStore.FindByIdAsync(message.WorkflowInstanceId);
-
-                if (!ValidatePreconditions(workflowInstanceId, workflowInstance, message.ActivityId))
-                    return;
-
-                await _workflowRunner.ResumeWorkflowAsync(
-                    workflowInstance!,
-                    message.ActivityId,
-                    message.Input);
+                await _mediator.Send(message);
             }
             finally
             {
@@ -67,35 +54,6 @@ namespace Elsa.Dispatch.Consumers
                 _stopwatch.Stop();
                 _logger.LogDebug("Held lock on workflow instance {WorkflowInstanceId} for {ElapsedTime}", workflowInstanceId, _stopwatch.Elapsed);
             }
-        }
-
-        private bool ValidatePreconditions(string? workflowInstanceId, WorkflowInstance? workflowInstance, string? activityId)
-        {
-            if (workflowInstance == null)
-            {
-                _logger.LogWarning("Could not run workflow instance with ID {WorkflowInstanceId} because it does not exist", workflowInstanceId);
-                return false;
-            }
-
-            if (workflowInstance.WorkflowStatus != WorkflowStatus.Suspended && workflowInstance.WorkflowStatus != WorkflowStatus.Running)
-            {
-                _logger.LogWarning("Could not run workflow instance with ID {WorkflowInstanceId} because it has a status other than Suspended or Running. Its actual status is {WorkflowStatus}", workflowInstanceId, workflowInstance.WorkflowStatus);
-                return false;
-            }
-
-            if (activityId != null)
-            {
-                var activityIsBlocking = workflowInstance.BlockingActivities.Any(x => x.ActivityId == activityId);
-                var activityIsScheduled = workflowInstance.ScheduledActivities.Any(x => x.ActivityId == activityId) || workflowInstance.CurrentActivity?.ActivityId == activityId;
-                
-                if (!activityIsBlocking && !activityIsScheduled)
-                {
-                    _logger.LogWarning("Did not run workflow {WorkflowInstanceId} for activity {ActivityId} because the workflow is not blocked on that activity nor is that activity scheduled for execution", workflowInstanceId, activityId);
-                    return false;
-                }
-            }
-
-            return true;
         }
     }
 }
