+++ src/server/Elsa.Server.Hangfire/Jobs/WorkflowDefinitionJob.cs
@@ -1,36 +1,14 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using Elsa.Dispatch;
-using Elsa.Models;
-using Elsa.Services;
-using Microsoft.Extensions.Logging;
+using MediatR;
 
 namespace Elsa.Server.Hangfire.Jobs
 {
     public class WorkflowDefinitionJob
     {
-        private readonly IWorkflowRegistry _workflowRegistry;
-        private readonly IStartsWorkflow _startsWorkflow;
-        private readonly ILogger<WorkflowDefinitionJob> _logger;
-
-        public WorkflowDefinitionJob(IWorkflowRegistry workflowRegistry, IStartsWorkflow startsWorkflow, ILogger<WorkflowDefinitionJob> logger)
-        {
-            _workflowRegistry = workflowRegistry;
-            _startsWorkflow = startsWorkflow;
-            _logger = logger;
-        }
-        
-        public async Task ExecuteAsync(ExecuteWorkflowDefinitionRequest request, CancellationToken cancellationToken = default)
-        {
-            var workflowBlueprint = await _workflowRegistry.GetWorkflowAsync(request.WorkflowDefinitionId, request.TenantId, VersionOptions.Published, cancellationToken);
-
-            if (workflowBlueprint == null)
-            {
-                _logger.LogWarning("No published workflow definition {WorkflowDefinitionId} found", request.WorkflowDefinitionId);
-                return;
-            }
-
-            await _startsWorkflow.StartWorkflowAsync(workflowBlueprint, request.ActivityId, request.Input, request.CorrelationId, request.ContextId, cancellationToken);
-        }
+        private readonly IMediator _mediator;
+        public WorkflowDefinitionJob(IMediator mediator) => _mediator = mediator;
+        public async Task ExecuteAsync(ExecuteWorkflowDefinitionRequest request, CancellationToken cancellationToken = default) => await _mediator.Send(request, cancellationToken);
     }
 }
