+++ src/server/Elsa.Server.Orleans/Grains/WorkflowInstanceGrain.cs
@@ -1,10 +1,8 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using Elsa.Dispatch;
-using Elsa.Persistence;
 using Elsa.Server.Orleans.Grains.Contracts;
-using Elsa.Services;
-using Microsoft.Extensions.Logging;
+using MediatR;
 using Orleans;
 using Orleans.Concurrency;
 
@@ -13,28 +11,8 @@ namespace Elsa.Server.Orleans.Grains
     [StatelessWorker]
     public class WorkflowInstanceGrain : Grain, IWorkflowInstanceGrain
     {
-        private readonly IWorkflowInstanceStore _store;
-        private readonly IResumesWorkflow _workflowRunner;
-        private readonly ILogger<WorkflowInstanceGrain> _logger;
-
-        public WorkflowInstanceGrain(IWorkflowInstanceStore store, IResumesWorkflow workflowRunner, ILogger<WorkflowInstanceGrain> logger)
-        {
-            _store = store;
-            _workflowRunner = workflowRunner;
-            _logger = logger;
-        }
-        
-        public async Task ExecuteWorkflowAsync(ExecuteWorkflowInstanceRequest request, CancellationToken cancellationToken = default)
-        {
-            var workflowInstance = await _store.FindByIdAsync(request.WorkflowInstanceId, cancellationToken);
-            
-            if(workflowInstance == null)
-            {
-                _logger.LogWarning("Workflow instance {WorkflowInstanceId} not found", request.WorkflowInstanceId);
-                return;
-            }
-            
-            await _workflowRunner.ResumeWorkflowAsync(workflowInstance, request.ActivityId, request.Input, cancellationToken);
-        }
+        private readonly IMediator _mediator;
+        public WorkflowInstanceGrain(IMediator mediator) => _mediator = mediator;
+        public async Task ExecuteWorkflowAsync(ExecuteWorkflowInstanceRequest request, CancellationToken cancellationToken = default) => await _mediator.Send(request, cancellationToken);
     }
 }
